+++ src/main/java/com/mraof/minestuck/data/worldgen/MSStructureProvider.java
@@ -59,9 +59,10 @@ public final class MSStructureProvider
 				Map.of(), GenerationStep.Decoration.SURFACE_STRUCTURES, TerrainAdjustment.NONE)));
 		context.register(SMALL_RUIN, new SmallRuinStructure(new Structure.StructureSettings(biomes.getOrThrow(MSTags.Biomes.HAS_SMALL_RUIN),
 				Map.of(), GenerationStep.Decoration.SURFACE_STRUCTURES, TerrainAdjustment.NONE)));
-		context.register(PROSPIT_BUNKER, jigsaw(biomes, pools, MSTags.Biomes.HAS_PROSPIT_BUNKER, minestuckMobCategoriesSpawnOverride, PROSPIT_BUNKER_START_POOL));
-		context.register(DERSE_BUNKER, jigsaw(biomes, pools, MSTags.Biomes.HAS_DERSE_BUNKER, minestuckMobCategoriesSpawnOverride, DERSE_BUNKER_START_POOL));
-		context.register(IMP_BUNKER, jigsaw(biomes, pools, MSTags.Biomes.HAS_IMP_BUNKER, consortSpawnOverride, IMP_BUNKER_START_POOL));
+		context.register(ARENA, jigsaw(biomes, pools, MSTags.Biomes.HAS_ARENA, TerrainAdjustment.BEARD_THIN, minestuckMobCategoriesSpawnOverride, ARENA_START_POOL));
+		context.register(PROSPIT_BUNKER, jigsaw(biomes, pools, MSTags.Biomes.HAS_PROSPIT_BUNKER, TerrainAdjustment.NONE, minestuckMobCategoriesSpawnOverride, PROSPIT_BUNKER_START_POOL));
+		context.register(DERSE_BUNKER, jigsaw(biomes, pools, MSTags.Biomes.HAS_DERSE_BUNKER, TerrainAdjustment.NONE, minestuckMobCategoriesSpawnOverride, DERSE_BUNKER_START_POOL));
+		context.register(IMP_BUNKER, jigsaw(biomes, pools, MSTags.Biomes.HAS_IMP_BUNKER, TerrainAdjustment.NONE, consortSpawnOverride, IMP_BUNKER_START_POOL));
 		context.register(ImpDungeon.KEY, new ImpDungeonStructure(new Structure.StructureSettings(biomes.getOrThrow(MSTags.Biomes.HAS_IMP_DUNGEON),
 				consortSpawnOverride, GenerationStep.Decoration.SURFACE_STRUCTURES, TerrainAdjustment.NONE)));
 		context.register(ConsortVillage.KEY, new ConsortVillageStructure(new Structure.StructureSettings(biomes.getOrThrow(MSTags.Biomes.HAS_CONSORT_VILLAGE),
@@ -102,12 +103,12 @@ public final class MSStructureProvider
 				new RandomSpreadStructurePlacement(40, 10, RandomSpreadType.LINEAR, 1481098009)));
 	}
 	
-	private static JigsawStructure jigsaw(HolderGetter<Biome> biomes, HolderGetter<StructureTemplatePool> templatePools, TagKey<Biome> biome, Map<MobCategory, StructureSpawnOverride> spawnOverrides, ResourceKey<StructureTemplatePool> startPool)
+	private static JigsawStructure jigsaw(HolderGetter<Biome> biomes, HolderGetter<StructureTemplatePool> templatePools, TagKey<Biome> biome, TerrainAdjustment terrainAdjustment, Map<MobCategory, StructureSpawnOverride> spawnOverrides, ResourceKey<StructureTemplatePool> startPool)
 	{
 		return new JigsawStructure(
 				new Structure.StructureSettings.Builder(biomes.getOrThrow(biome))
 						.generationStep(GenerationStep.Decoration.SURFACE_STRUCTURES)
-						.terrainAdapation(TerrainAdjustment.NONE)
+						.terrainAdapation(terrainAdjustment)
 						.spawnOverrides(spawnOverrides)
 						.build(),
 				templatePools.getOrThrow(startPool),