+++ src/main/java/com/mraof/minestuck/item/loot/MSLootTables.java
@@ -26,6 +26,11 @@ public class MSLootTables
 	public static final ResourceKey<LootTable> FROG_TEMPLE_CHEST = resourceKey("chests/frog_temple");
 	public static final ResourceKey<LootTable> BASIC_MEDIUM_CHEST = resourceKey("chests/medium_basic");
 	public static final ResourceKey<LootTable> MEDIUM_SUPPLY_CHEST = resourceKey("chests/medium_supply");
+	public static final ResourceKey<LootTable> MEDIUM_TRIAL = resourceKey("chests/medium_trial");
+	public static final ResourceKey<LootTable> MEDIUM_TRIAL_KEY = resourceKey("chests/medium_trial_key");
+	public static final ResourceKey<LootTable> MEDIUM_VAULT = resourceKey("chests/medium_vault");
+	public static final ResourceKey<LootTable> MEDIUM_OMINOUS_VAULT = resourceKey("chests/medium_ominous_vault");
+	public static final ResourceKey<LootTable> REGIWEAPONS = resourceKey("chests/regiweapons");
 	public static final ResourceKey<LootTable> CONSORT_JUNK_REWARD = resourceKey("gameplay/consort_junk");
 	public static final ResourceKey<LootTable> CONSORT_FOOD_STOCK = resourceKey("gameplay/consort_food");
 	public static final ResourceKey<LootTable> CONSORT_GENERAL_STOCK = resourceKey("gameplay/consort_general");